backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pinecone-client
google-generativeai
pydantic
pydantic-settings
python-multipart
httpx
numpy
python-dotenv
